Title: 최단 왕복, Time: 85ms, Memory: 9MB, Status: Runtime Error - CodeTree

+++ 241204/최단 왕복/shortest-round-trip.java
@@ -21,7 +21,7 @@ public class Main {
 
             int x = Integer.parseInt(st.nextToken());
             int y = Integer.parseInt(st.nextToken());
-            graph[x][y] = Integer.parseInt(st.nextToken());
+            graph[x][y] = Long.parseLong(st.nextToken());
 
             graph[i][i] = 0;
         }
